Moved some identity related messages to common and added a message for policy.
